No longer use MaxBytesReader for determining file size
The main reason we don't want to use MaxBytesReader is because we don't
know how far down the chain of potential middleware the request could be
before validation, and whether or not a previous ParseMultipartForm call
has been made.

For multipart/form-data requests we will simply check the Size property
on the returned multipart.FileHeader struct, and for every other request
we will check the number of bytes read into the buffer.

This proves to be a more reliable method for checking the file size
being uploaded.