Update to v2
Introduced some breaking changes around how files are unmarshalled from
requests and how form validation now works, so bumping this to v2.

The Validator interface has been removed and the Form interface has been
updated to include a Validate method that takes a context.Context as its
only argument and returns an error. In retrospect it makes more sense
for this to be on the Form since this is what is being validated.

The File type has been reduced in scope, it now simply wraps the
multipart.File type, and has the Header and Type fields for the file
header and mime type respectively.

The UnmarshalFile and UnmarshalFiles functions can be used for
retrieving files from a given request. The UnmarshalFiles function will
return every file uploaded in a multipart request. The UnmarshalFile
function supports parsing the entire request body for a file upload, if
the Content-Type is not of multipart/form-data. This will also support
writing the file to disk should it exceed the size of 32MB.

The FieldError type has been added to capture any errors that may occur
for a form field during validation. This would wrap the underlying error
that caused the validation to fail.